refactor(envoy): legacy check-run ids decode on the id type, not a State mirror; resync pages every rollup for the fence
The ordering fence needs the highest check-run id on the head, which can
sit on a later contexts page even when the rollup is green; failing names
still read only when failing. The legacy string id decode moves onto a
checkRunID type so State keeps the alias shim and a new field cannot
silently vanish on restart.

Omp-Session: 01a06999-b559-7000-8f0f-b9341ba4bb1e

+++ packages/envoy/internal/cistore/render.go
@@ -86,8 +86,8 @@ func renderSummary(s State) Summary {
 	failingChecks := make([]FailingCheck, 0)
 	var latestCheckRunID uint64
 	for key, check := range s.Checks {
-		if check.CheckRunID > latestCheckRunID {
-			latestCheckRunID = check.CheckRunID
+		if uint64(check.CheckRunID) > latestCheckRunID {
+			latestCheckRunID = uint64(check.CheckRunID)
 		}
 		name := check.Name
 		if name == "" {
